Allow runtime extension of Java types
This change introduces two new internal annotations, which can be used
to signal that a type (or property) should be extended at runtime by
a domain-specific interface. The use case is for entity and typelist
types, which have historically been instances of IEntityType and
ITypeList at runtime, but which otherwise would not be without the
entity typeloader.

If a type is annotated with ExtendedType, then, when constructing the
JavaType for that type, we consult the IEntityAccess to ask for a
factory to produce the extension data for that type. If something is
returned, then we wrap the JavaType in a proxy that implements both
IJavaType(Internal), as well as the additional interface indicated by
the factory.

The same is done for Java properties.

Because some Java types and properties are now proxied, we must be sure
to downcast to the interfaces IJavaType and IJavaPropertyInfo, rather
than the classes, JavaType and JavaPropertyInfo.