PL-35140.  Support a call-site that otherwise results in IllegalAccessExpression due to accessing a method/field in a class in the same package, but loaded in a different class loader from the call-site.  Generally the change involves modifying the Gosu compiler to detect if the call-site could be a making an illegal call and, instead of generating a normal invokeXxx or getfield bytecode, generate a method call or field access reflectively.   It is important to note that not only do we need to generate the call reflectively, but also we must also maintain a reference to the method-call or field-access statically.  This is because the Studio incremental compiler relies on Jetbrains internal code to determine the dependency graph.  This code analyzes bytecode to determine the graph, hence the need for the static references.
    The original idea was to generate the code in a try-catch statement like this:
    var temp: <result-type>
    try {
      temp = x.foo(...)
    }
    catch( e: IllegalAccessError ) {
      temp = callReflectively( x, "foo", ... )
    }
    temp

However, this proved fatal because at the entry point after the caught Error in the catch-clause the JVM has cleared the stack.  Thus, if we have a call like so:

    field1 = foo(...)

The lhs of the assignment (field1) will no longer be on the stack and the bytecode will not work, let alone verify.

So we opted to tighten up the static conditions for genetating a reflective call, we are much smarter about that now, and we generate a simple if-statement instead of the try-catch:

    var temp: <result-type>
    if( false ) {
      temp = x.foo(...)
    }
    else {
      temp = callReflectively( x, "foo", ... )
    }
    temp

The if( false ) part of the statement there only to please Studio's incremental compiler; the call is never made statically.  But, again, we've tightened up the conditions under which this code is generated -- reflective calls happen less frequently than originally anticipated.