Add new property syntax.  Basically the new syntax is similar to declaring a field using 'var', but instead you use the 'property' keyword:
class MyClass {
  property Foo: String
}

The Foo property is readable and writable; accessor methods are created internally.  The property is backed by the hidden private field, _Foo, which can be accessed inside the class.  Effectively this is the same as writing:

class MyClass {
  var _Foo: String

  property get Foo(): String {
    return _Foo
  }
  property set Foo( value: String ) {
    _Foo = value
  }
}

When in an interface, the property is abstract:

interface IFace {
  property Foo: String
}

This has the same effect as writing this:

interface IFace {
  property get Foo(): String
  property set Foo( value: String )
}

You can also declare a default value:

class MyClass  {
  property Foo: String = "hi"
}

Which initializes the backing field to "hi":

var _Foo: String = "hi"

You can declare read-only and write-only properties:

class MyClass {
  property get Foo: String
}

Which is the same as this:

class MyClass {
  var _Foo: String

  property get Foo(): String {
    return _Foo
  }
}

You can easily declare and implement properties:

 interface IFace {
   property Foo: String
 }

 class ImplClass implements IFoo {
   override property Foo: String
 }

 You can do justs about anything that makes sense with this new syntax.

 Note this syntax is meant to replace the shorthand syntax bolted on to field declarations.  E.g.,

 class MyClass {
   var _foo: String as Foo = "hi"
 }

Should be written with the new syntax as:

class MyClass {
  property Foo: String = "hi"
}

The new syntax is easier to use and understand, more consistent with existing longhand property syntax, and fixes some issues that are difficult or impossible to fix regarding the var shorthand syntax e.g., using the override keyword, where to apply annotations, abstract declarations, etc.