Gosu Core - Type inference -- Support out-of-order type inference.  For instance, given a function call to generic method foo(),  the types inferred in an argument to foo can be inferred from inferences made in subsequent arguments:
var x = foo( \ a ->a.Alpha, \ b ->b.Code, MyEnum.First )
function foo<B, A>( cb(A), func(B): A, ref: B ) : A { return null }
static enum MyEnum { First, Second }

The call to foo() involves out of order inference since the first arg relies on the second arg's inference of A and the second arg relies on the third arg's inference of B, thus the type information for the call requires three passes of analysis.  Note if we reverse the order of foo's parameters, the type information requires only one pass.