https://guidewirejira.atlassian.net/browse/ISGOSU-264
The main defect: The cache uses an unmanaged ReferenceQueue to create WeakReferences, which has the effect of 1. prematurely releasing type references and 2. retaining all the WeakReference wrappers in the heap. Over time the WeakReferences build up and eventually consume the heap.  Perhaps worse is the performance overhead introduced by having to dynamically recompile types after they are prematurely released.  The fix involves:

- Change WeakReference usage to SoftReference.  Retains type references until process encounters max heap threshold, which should not happen on a properly configured node. Also improves performance by avoiding unnecessary dynamic recompilation of type references.
- Remove use of ReferenceQueue.  Avoids retaining reference wrappers.
- Remove WeakReference subclass.  The extra class is unneeded since its sole purpose was to manage the reference queue.